chore: separate ui and Chart modules
