+++ src/utilities/codeEditorUtils.js
@@ -6,6 +6,8 @@ const path = require('path');
 const os = require('os');
 const fs = require('fs');
 const { spawnSync } = require('child_process');
+import { EEChart } from "./eeChart";
+import { EEUi } from "./eeUi";
 
 function tryGitCommand(repoPath, gitCommand) {
     try {
@@ -290,8 +292,8 @@ class Tools {
         this.Export = new ExportConstructor(ee, successCallback, errCallback);
 
         /* These are defined to be silently ignored*/
-        this.ui = ui;
-        this.Chart = Chart;
+        this.ui = EEUi;
+        this.Chart = EEChart;
     }
 
     Log = function(log){
@@ -456,414 +458,3 @@ class Tools {
     }
 }
 exports.Tools = Tools;
-
-/* Mock implementation of ui, so that
-   they can be ignored in the user code without raising
-   errors. 
-   Some of this could be implemented to work with the 
-   leaflet map..
-*/
-const ui = {
-  Button: function(){
-    return {
-      getDisabled:function(){},
-      getImageUrl:function(){},
-      getLabel:function(){},
-      onClick:function(){},
-      setDisabled:function(){},
-      setImageUrl:function(){},
-      setLabel:function(){},
-      style:function(){},
-      unlisten:function(){},
-    };
-  },
-  Chart: function () {
-    return {
-      getChartType: function(){},
-      getDataTable:function(){},
-      getDownloadable:function(){},
-      getOptions:function(){},
-      getView:function(){},
-      onClick:function(){},
-      setChartType:function(){},
-      setDataTable:function(){},
-      setDownloadable:function(){},
-      setOptions:function(){},
-      setSeriesNames:function(){},
-      setView:function(){},
-      style:function(){},
-      unlisten:function(){},
-    };
-  },
-  Checkbox: function(){
-    return {
-        getDisabled:function(){},
-        getLabel:function(){},
-        getValue:function(){},
-        onChange:function(){},
-        setDisabled:function(){},
-        setLabel:function(){},
-        setValue:function(){},
-        style:function(){},
-        unlisten:function(){},       
-    };
-  },
-  DateSlider: function(){
-    return {
-        getDisabled:function(){},
-        getEnd:function(){},
-        getPeriod:function(){},
-        getStart:function(){},
-        getValue:function(){},
-        onChange:function(){},
-        setDisabled:function(){},
-        setEnd:function(){},
-        setPeriod:function(){},
-        setStart:function(){},
-        setValue:function(){},
-        style:function(){},
-        unlisten:function(){},
-    };
-  },
-  Label: function(){
-    return {
-        getImageUrl:function(){},
-        getUrl:function(){},
-        getValue:function(){},
-        setImageUrl:function(){},
-        setUrl:function(){},
-        setValue:function(){},
-        style:function(){},
-    };
-  },
-  Map: function(){
-    return {
-        add:function(){},
-        addLayer:function(){},
-        centerObject:function(){},
-        clear:function(){},
-        drawingTools:function(){},
-        getBounds:function(){},
-        getCenter:function(){},
-        getScale:function(){},
-        getZoom:function(){},
-        insert:function(){},
-        layers:function(){},
-        onChangeBounds:function(){},
-        onChangeCenter:function(){},
-        onChangeZoom:function(){},
-        onClick:function(){},
-        onIdle:function(){},
-        onTileLoaded:function(){},
-        remove:function(){},
-        setCenter:function(){},
-        setControlVisibility:function(){},
-        setGestureHandling:function(){},
-        setLocked:function(){},
-        setOptions:function(){},
-        setZoom:function(){},
-        style:function(){},
-        unlisten:function(){},
-        widgets:function(){},
-    };
-  },
-  Panel: function(){
-    return {
-        add:function(){},
-        clear:function(){},
-        getLayout:function(){},
-        insert:function(){},
-        remove:function(){},
-        setLayout:function(){},
-        style:function(){},
-        widgets:function(){},
-    };
-  },
-  Select: function(){
-    return{
-        getDisabled:function(){},
-        getPlaceholder:function(){},
-        getValue:function(){},
-        items:function(){},
-        onChange:function(){},
-        setDisabled:function(){},
-        setPlaceholder:function(){},
-        setValue:function(){},
-        style:function(){},
-        unlisten:function(){},
-    };
-  },
-  Slider: function(){
-    return{
-        getDisabled:function(){},
-        getMax:function(){},
-        getMin:function(){},
-        getStep:function(){},
-        getValue:function(){},
-        onChange:function(){},
-        onSlide:function(){},
-        setDisabled:function(){},
-        setMax:function(){},
-        setMin:function(){},
-        setStep:function(){},
-        setValue:function(){},
-        style:function(){},
-        unlisten:function(){},
-    };
-  },
-  SplitPanel: function(){
-    return{
-        getFirstPanel:function(){},
-        getOrientation:function(){},
-        getPanel:function(){},
-        getSecondPanel:function(){},
-        getWipe:function(){},
-        setFirstPanel:function(){},
-        setOrientation:function(){},
-        setPanel:function(){},
-        setSecondPanel:function(){},
-        setWipe:function(){},
-        style:function(){},
-        unlisten:function(){},
-    };
-  },
-  Textbox: function(){
-    return{
-        getDisabled:function(){},
-        getPlaceholder:function(){},
-        getValue:function(){},
-        onChange:function(){},
-        setDisabled:function(){},
-        setPlaceholder:function(){},
-        setValue:function(){},
-        style:function(){},
-        unlisten:function(){},
-    };
-  },
-  Thumbnail: function(){
-    return{
-        getImage:function(){},
-        getParams:function(){},
-        onClick:function(){},
-        setImage:function(){},
-        setParams:function(){},
-        style:function(){},
-        unlisten:function(){},
-    };
-  },
-};
-
-ui.Chart.array = {
-    values: function(){return ui.Chart();},
-};
-ui.Chart.feature = {
-  byFeature: function () {return ui.Chart();},
-  byProperty: function () {return ui.Chart();},
-  groups: function () {return ui.Chart();},
-  histogram: function () {return ui.Chart();},
-};
-ui.Chart.image = {
-  byClass: function () {return ui.Chart();},
-  byRegion: function () {return ui.Chart();},
-  doySeries: function () {return ui.Chart();},
-  doySeriesByRegion: function () {return ui.Chart();},
-  histogram: function () {return ui.Chart();},
-  regions: function () {return ui.Chart();},
-  series: function () {return ui.Chart();},
-  seriesByRegion: function () {return ui.Chart();},
-};
-ui.Map.CloudStorageLayer = function(){
-    const m = {}; 
-    m.getBucket = function(){};
-    m.getBucket=function(){};
-    m.getMaxZoom=function(){};
-    m.getName=function(){};
-    m.getOpacity=function(){};
-    m.getPath=function(){};
-    m.getShown=function(){};
-    m.getSuffix=function(){};
-    m.setBucket=function(){};
-    m.setMaxZoom=function(){};
-    m.setName=function(){};
-    m.setOpacity=function(){};
-    m.setPath=function(){};
-    m.setShown=function(){};
-    m.setSuffix=function(){};
-    return m;
-};
-ui.Map.DrawingTools = function(){
-    const m = {};
-    m.addLayer=function(){};
-    m.clear=function(){};
-    m.draw=function(){};
-    m.edit=function(){};
-    m.get=function(){};
-    m.getDrawModes=function(){};
-    m.getLinked=function(){};
-    m.getMap=function(){};
-    m.getSelected=function(){};
-    m.getShape=function(){};
-    m.getShown=function(){};
-    m.layers=function(){};
-    m.onDraw=function(){};
-    m.onEdit=function(){};
-    m.onErase=function(){};
-    m.onLayerAdd=function(){};
-    m.onLayerConfig=function(){};
-    m.onLayerRemove=function(){};
-    m.onLayerSelect=function(){};
-    m.onSelect=function(){};
-    m.onShapeChange=function(){};
-    m.set=function(){};
-    m.setDrawModes=function(){};
-    m.setLinked=function(){};
-    m.setSelected=function(){};
-    m.setShape=function(){};
-    m.setShown=function(){};
-    m.stop=function(){};
-    m.toFeatureCollection=function(){};
-    m.unlisten=function(){};
-};
-ui.Map.FeatureViewLayer = function(){
-    const m = {};
-    m.getAssetId=function(){};
-    m.getName=function(){};
-    m.getOpacity=function(){};
-    m.getShown=function(){};
-    m.getVisParams=function(){};
-    m.setAssetId=function(){};
-    m.setName=function(){};
-    m.setOpacity=function(){};
-    m.setShown=function(){};
-    m.setVisParams=function(){};
-};
-ui.Map.GeometryLayer = function(){
-    const m = {};
-    m.fromGeometry=function(){};
-    m.geometries=function(){};
-    m.get=function(key){};
-    m.getColor=function(){};
-    m.getEeObject=function(){};
-    m.getLocked=function(){};
-    m.getName=function(){};
-    m.getShown=function(){};
-    m.openConfigurationDialog=function(){};
-    m.set=function(){};
-    m.setColor=function(){};
-    m.setLocked=function(){};
-    m.setName=function(){};
-    m.setShown=function(){};
-    m.toGeometry=function(){};
-};
-ui.Map.Layer = function() {
-    const m = {};
-    m.getEeObject=function(){};
-    m.getName=function(){};
-    m.getOpacity=function(){};
-    m.getShown=function(){};
-    m.getVisParams=function(){};
-    m.setEeObject=function(){};
-    m.setName=function(){};
-    m.setOpacity=function(){};
-    m.setShown=function(){};
-    m.setVisParams=function(){};
-};
-ui.Map.Linker = function(){
-    const m = {};
-    m.add=function(){};
-    m.forEach=function(){};
-    m.get=function(){};
-    m.getJsArray=function(){};
-    m.insert=function(){};
-    m.length=function(){};
-    m.remove=function(){};
-    m.reset=function(){};
-    m.set=function(){};
-};
-
-ui.Panel.Layout = {
-    absolute:function(){},
-    flow:function(){},
-};
-ui.data = {
-    ActiveDictionary:function(){
-    return {
-        get: function(){},
-        set: function(){}
-    };
-    },
-    ActiveList:function(){
-    return {
-        add:function(){},
-        forEach:function(){},
-        get:function(){},
-        getJsArray:function(){},
-        insert:function(){},
-        length:function(){},
-        remove:function(){},
-        reset:function(){},
-        set:function(){},
-    };
-    },  
-};
-ui.root = {
-    add:function(){},
-    clear:function(){},
-    getLayout:function(){},
-    insert:function(){},
-    onResize:function(){},
-    remove:function(){},
-    setKeyHandler:function(){},
-    setLayout:function(){},
-    widgets:function(){},
-}
-ui.url = {
-    get:function(){},
-    set:function(){},
-}
-ui.util = {
-    clear:function(){},
-    clearTimeout:function(){},
-    debounce:function(){},
-    getCurrentPosition:function(){},
-    rateLimit:function(){},
-    setInterval:function(){},
-    setTimeout:function(){},
-    throttle:function(){},
-}
-ui.Key={};
-
-/*
-    Mock implementation of Chart
-    Note that Chart is now deprecated
-    (we could take it from the ui.Chart.. )
-*/
-const Chart = function(){
-    return {
-        setChartType:function(){},
-        setDataTable:function(){},
-        setOptions:function(){},
-        setSeriesNames:function(){},
-        setView:function(){},
-        transform:function(){},
-    };
-};
-Chart.array = {
-    values: function(){return Chart();},
-};
-Chart.feature = {
-    byFeature: function(){return Chart();},
-    byProperty: function(){return Chart();},
-    groups: function(){return Chart();},
-    histogram: function(){return Chart();},
-};
-Chart.image = {
-    byClass: function(){return Chart();},
-    byRegion: function(){return Chart();},
-    doySeries: function(){return Chart();},
-    doySeriesByRegion: function(){return Chart();},
-    histogram: function(){return Chart();},
-    regions: function(){return Chart();},
-    series: function(){return Chart();},
-    seriesByRegion: function(){return Chart();},
-};
